Fix TypeError in counters
In your plusFifty method, you have ctx: commands.Context as the first parameter. Because plusFifty is a method of the Counters class, Python is interpreting ctx as the self. (Which is why we define it and add it :3)

Consequently, when the command is invoked, the twitchio.ext.commands.Context object is passed to the counted parameter. (That's fine, though! It's a one line fix :3)

You are attempting to perform counted += 1, which translates to Context_object + 1, which is an operation that isn't supported. This should resolve it as we're now passing self so ctx won't be attributed. ^^

For uniqueID I changed the logic to len(counters) + 1, because we're essentially passing the entire counters list as the ID. This change makes it a bit more conventional <3